Migrate realms to new sourcing library (#3356)

+++ delta/app/src/test/scala/ch/epfl/bluebrain/nexus/delta/utils/RouteFixtures.scala
@@ -191,7 +191,7 @@ trait RouteFixtures extends TestHelpers with IOValues {
 
   def realmMetadata(
       label: Label,
-      rev: Long = 1L,
+      rev: Int = 1,
       deprecated: Boolean = false,
       createdBy: Subject = Anonymous,
       updatedBy: Subject = Anonymous
